test: properly close mongodb/mongoose connections
This fixes the warning from jest and the test not being finished:

Jest did not exit one second after the test run has completed.
This usually means that there are asynchronous operations that weren't stopped in your tests. Consider running Jest with `--detectOpenHandles` to troubleshoot this issue.

The mongoose test file still causes this, not sure why